Create a tree-shakable module for realtime publishing
Owen mentioned that we have many browser use cases which only require
subscriptions, and no publishing. He suggested that we create a separate
tree-shakable module for this functionality.

This commit introduces the API, but the bundle size savings are minimal
since it only pulls out the very low-hanging fruit. I think that we
could return to this at some point to see what further size savings we
could achieve, but I didn’t want to spend too much time on this now.

Resolves #1491.