Make RequireJS work in modules.test.js
This gives these tests access to the same test helper APIs that the
other ones use. And, it means that we can access test helpers inside the
body of a `describe` block, which I’ll need in an upcoming commit in
order to call the shared delta tests added in 09354d4.