Remove the Types namespace
We’re removing it because we believe that it’s more consistent with
other JS libraries for users to write `import * as Ably from 'ably'` and
then access types on the `Ably` namespace.

The triggering of @typescript-eslint/no-redeclare (and hence the added
eslint-disable-next-line statements) seems to be related to [1]. I’m not
sure why it wasn’t complaining before, though.

Resolves #909.

[1] https://github.com/typescript-eslint/typescript-eslint/issues/2818