Rename RealtimePresence module to RealtimePresenceModule
To avoid a clash with the Types.RealtimePresence class (which describes
the realtime presence API) when we remove the Types namespace in #909.

This is ugly; it’s not consistent with the other modules, which don’t
have "Module" in their name. I’d welcome suggestions of alternative
things I could do here.