Correct tsc module settings for test:package
I noticed that, when I tried changing the structure of our package to
put the type declarations in a subdirectory (not part of this commit),
test:package started failing, claiming that `ably/modules` didn’t exist.

Changing the module settings per [1] fixes this, and sounds like it’s
the way it should have been configured in the first place:

> Use `esnext` with `--moduleResolution bundler` for bundlers, Bun, and
> tsx.

(Not sure _why_ putting the types in a subdirectory triggered this…)

I noticed that making this change caused VS Code to start displaying the
following error against the Ably import in `index-default.ts` (although
weirdly, not causing `tsc` to fail):

> Could not find a declaration file for module 'ably'.
> '/Users/lawrence/code/work/ably/ably-js/test/package/browser/build/node_modules/ably/build/ably.js'
> implicitly has an 'any' type.
>
> There are types at
> '/Users/lawrence/code/work/ably/ably-js/test/package/browser/build/node_modules/ably/ably.d.ts',
> but this result could not be resolved when respecting package.json
> "exports". The 'ably' library may need to update its package.json or
> typings.ts (7016)

So I’ve added typings against the "." package.json export. I think I
should have done this in de5ddfa.

[1] https://www.typescriptlang.org/docs/handbook/modules/reference.html#summary-1