Strip less-important logging from modular variant
In order to reduce bundle size further, we’ve decided to strip all
logging from the modular variant of the SDK, except for errors and
certain network events. (We also considered providing a separate
tree-shakable module with all of this logging code so that a user of the
modular variant of the library can opt in to it, but decided against it
for now; we might add it in later. This does mean that there is
currently no version of the SDK that allows you to use both deltas and
verbose logging on web.)

I couldn’t find any out-of-the-box esbuild functionality that let us do
this. The only stuff I could find related to stripping code was:

- the `pure` option, but that code only gets stripped if you minify the
  code (and even in that case I couldn’t actually get it to be stripped,
  perhaps would have been able to with further trying though), but
  minifying our generated modules bundle causes the bundle size of those
  who use it (as tested by our modulereport script) to increase
  considerably (for reasons I’m not sure of)

- the `drop` option, but that only lets you remove calls to `console` or
  `debugger`

So instead I’ve implemented it as an esbuild plugin.

Resolves #1526.