Emit JUnit report from tests
The test reports emitted by mocha-junit-reporter have <testcase> tags
with slightly weird `name` and `classname` properties, e.g.

<testsuite name="rest/defaults">
    <testcase name="rest/defaults Init with given environment" classname="Init with given environment">
    <testcase/>
</testsuite>

This looks a bit odd in the observability server web UI, which displays
the testcase.name and testcase.classname properties. I don’t know
whether it’d be better to do some pre-upload manipulation of the ably-js
reports, or to change the fields displayed in the web UI. Not going to
do anything about it now; it’s usable enough and I don’t want to mess up
the display of the ably-cocoa test reports.

Co-authored-by: Owen Pearson <owen.pearson@ably.com>