Only export types from default variant of library
When we remove the Types namespace in #909, the RealtimePresence class
defined in ably.d.ts will clash with the variable of the same name
declared in modules.d.ts.

Owen indicated that his preference for resolving this issue is to export
the common types (i.e. those currently contained in the Types namespace)
from only the default variant and not from the modular variant, so
that’s what we do here. Users of the modular variant who wish to refer
to the common types will need to import them from the default variant.

To maintain TypeDoc’s ability to generate hyperlinks from the references
which are contained in the modular variant but which refer to the common
types (e.g. some @link tags, BaseRest’s inheritance from AbstractRest),
we switch to using a single TypeDoc project to generate documentation
for both variants.

For reasons that I don’t understand, TypeDoc is unable to resolve the
aforementioned @link tags unless I explicitly add a module source [1].
I’d have thought that it’d be sufficient to simply `import` the types
that these tags refer to, but doing so makes no difference. So, I’ve
added these module sources. Unfortunately, IntelliSense in VS Code
doesn’t know what to do with these and the link is not clickable. (I
also tried experimenting with TSDoc’s syntax for declaration references
[2] but it made no difference). I think we’ll have to live with this;
luckily there are only a few such tags.

[1] https://typedoc.org/guides/declaration-references/#module-source
[2] https://tsdoc.org/pages/tags/link/