Convert message `encode` to use promises
The changes to the `connectionQueuing` test are because the test assumed
that the message would get marked as `sendAttempted` synchronously as
soon as `publish` was called. This relied on the fact that when no
encryption is required, the Message `encode` function called its
callback synchronously. However, this was an implementation detail (this
function has an asynchronous signature), and one which is no longer true
now that promises are used.