Convert Resource request methods to use promises
The `throwError` parameter here is to handle the fact that, when using
an HttpPaginatedResponse, PaginatedResource wishes to have access to
both the error that comes from the HTTP client _and_ the response body,
before deciding how to handle an error (see the `returnErrOnly`
function).

I’ve converted Resource.do to return a promise here, but I’ll do the
full conversion of that method in a separate commit for readability’s
sake.

TODO explain mangling