Document IXHRRequest’s emitted events
Needed to understand them for something I was investigating, thought it
worth documenting my findings.

Also made the arg types consistent between those expected by
CometTransport and those emitted by the XHR implementations.

TODO still a bit wip — need to double check the params to all of the
on('complete') and on('data'), and also switch to IPartialErrorInfo