Log all HTTP requests
Owen requested this during a review of #1526, which strips most of our
logging from the modular variant of the SDK, but aims to still log all
network activity.

A bonus benefit is that we now also have logging for requests which get
tried against different hosts, which we previously didn’t.

The log statements are adaptations of the existing ones in resource.ts.