Remove duplicate type names
We were using TypeDoc’s declaration reference functionality [1] to
disambiguate these, but this broke when we upgraded TypeDoc in 5f69795 —
it started generating link text of ":type" instead of the name of the
type.

This is because TypeDoc 0.24 added the --useTsLinkResolution option,
which is on by default. This option parses @link tags using the
TypeScript compiler. It does this in order to more closely match the way
in which editors such as VS Code will handle @link tags. However, the
TypeScript compiler’s parsing does not support TypeDoc declaration
syntax (and does not offer its own alternative), only falling back to
TypeDoc @link parsing when TypeScript fails to parse a @link. (See the
TypeDoc issue I raised [2] for further context.)

So, stop using declaration references to disambiguate types.

We still make some use of declaration references in modules.d.ts, to
link to the types in ably.d.ts and to link to a specific overload of a
method. These are still being parsed correctly (because TypeScript is
unable to parse them) and there is no good alternative, so I’ll leave
them as is, accepting that editors won’t handle them well (as was always
the case).

Resolves #1560.

[1] https://typedoc.org/guides/declaration-references/
[2] https://github.com/TypeStrong/typedoc/issues/2485