Highlight that whenPromiseSettles doesn’t guarantee error type
When I added whenPromiseSettles in cd6c2d8, I didn’t realise that
TypeScript types the `catch` callback’s argument as `any`, which doesn’t
make much sense to me given that the argument of a `catch` clause is
typed as `unknown` (the latter seeming more appropriate).

So, make it explicit that we’re performing a type assertion on the
error.