Adds `exports."./react"` to package.json
Also removes "exports" property from package.react.json in react-hooks
folder that is copied to `react` build folder. This file now only has
`"main": "./cjs/index.js"`, which, for not obvious reason to me, needs
to be present in order for `npx attw` check to not show "Resolution
failed" error for node10 for "ably/react" path.

Resolves #1622