Fix package-lock.json optional dependency issue
`npm run test-ct` was failing in CI (see job log [1]) with reference
to a npm bug [2] and suggested to delete package-lock.json and do
`npm install` to rebuild it. This commit has the result of it.

Also this commit changes `lockfileVersion` in package-lock.json to 3,
which is backwards compatible to npm v7 (bundled with node starting from
version 15). We're using node16 in CI for package tests, so it's fine.

[1] https://github.com/ably/ably-js/actions/runs/7985041805/job/21802827410?pr=1625
[2] https://github.com/npm/cli/issues/4828