Bring whenState in line with spec and test it
Until spec commit d5e8fc6, these methods were exclusive to ably-js and
were not tested. This commit brings the ably-js implementations in line
with the aforementioned spec commit and tests that they conform to the
spec.

Documentation based on sdk-api-reference commit 4a06b46.

Note the addition of the async/await adaptations of closeAndFinish and
monitorConnection. These new tests are the first realtime tests that
we’ve written on the v2 branch — i.e. since we dropped support for the
callbacks API, and so I’ve written them using async/await properly
instead of the whenPromiseSettles stopgap solution of 43a2d1d. Hence the
new helper methods.