Remove dual promise/callback operation of whenState
Missed this in 7c465d6.

I’ve had to remove the assertion from channelattachWhenState which
checks that whenState calls its listener immediately, because there’s no
easy way to check for this (because promise listeners are always invoked
asynchronously and there’s no way to synchronously check whether a
promise is pending).