consumergroup: use modulo-based hashing scheme
The `hashring` package is node-only as it depends on the native `crypto`
package. Replaced with a simple modulo hash scheme for now.

Fixes the case where the channel is already attached and the channel is
obtained with new rewind options that require a re-attach.

Updates the consumer group partitioning test to more robustly assert
that channels are partitioned across consumers.