channelgroups: do not share channel object
The Channels object used by the ChannelGroup for internal channel
operations no longer shares the same object exposed on the client via
the .channels() method. This is to ensure that independent usage of an
individual channel that happens to be included in a channel group is not
impacted by its usage in the channel group.