Revert "Replace the existing plugins mechanism"
This reverts commit c1f70b0. In that commit, we removed the ability to
inject optional functionality into the default variant of the SDK for
browsers, as well as into all non-browser platforms. We have since
realised that we wish to maintain this ability; for example the Realtime
team is currently working on an experimental channel group feature which
we don’t wish to currently bundle with the SDK.

We’ve decided that the best way to provide this ability is to restore
the ClientOptions.plugins mechanism which we removed in c1f70b0.

Next, in order to still only have a single mechanism for injecting
functionality into the SDK, we’ll update the modular variant of the SDK
to receive its functionality via ClientOptions.plugins instead of a
separate constructor argument.

See related discussion in [1].

[1] https://ably.atlassian.net/wiki/spaces/SDK/pages/2930671626/Back+to+plugins+from+modules+option+for+Ably+JS+v2