Use ClientOptions.plugins for modular variant
This updates the modular variant of the SDK to receive its functionality
via the ClientOptions.plugins mechanism which we restored in 5f56c45.
The aim of this change is to make sure that there is still only a single
mechanism for injecting functionality into the SDK, as there was
pre-5f56c45.

As part of this change, I’ve also replaced the use of the term “module”,
when used to refer to a piece of injectable functionality, with the term
“plugin”. The term “module” already has a defined meaning in ECMAScript
(see [1]; a module is a file, basically), so we don’t want to overload
it.

This new usage of ClientOptions.plugins (i.e. plugins unique to a given
client library) is allowed for by spec point PC4, which I added in
commit 79f9bd0 in [2] (awaiting approval at time of writing).

Resolves #1662.

[1] https://tc39.es/ecma262/multipage/ecmascript-language-scripts-and-modules.html#prod-Module
[2] https://github.com/ably/specification/pull/184