channelgroup: unsubscribe channel after timeout
In order to avoid keeping the channel alive, we add a configurable
timeout after which the channel will be unsubscribed if no messages are
received. This is to avoid keeping the channel active. This can lead to
missed messages if the a message is published after the client
unsubscribes and before the channel becomes inactive. This is an
acceptable edge case for the client-side simulation, especially with the
default 1h timeout.