Add migration guide for v2
Based on e71ed5f...2e12459.

Note that 1.2.50, to which this guide refers, doesn't exist yet. We’ll
do that in #1672.

The description of the stats response changes comes from Simon in [1],
and the description of the removal of connection upgrades comes from
Owen in [2].

HTML tags used for headings so that I can specify a fixed `id` to use in
fragment links, because I’m probably going to keep editing the headings
and so if I rely on the generated headings then I’ll almost certainly
end up with a broken link. (GitHub doesn’t seem to support the custom
heading ID syntax described in [3].) And we don’t have a linter that
checks for broken fragment links.

Resolves ECO-1416.

[1] https://github.com/ably/ably-js/pull/1670#discussion_r1532093309
[2] https://github.com/ably/ably-js/pull/1670#discussion_r1532206072
[3] https://www.markdownguide.org/extended-syntax/#heading-ids