Remove incorrect deprecation of ClientOptions.maxMessageSize
This client option still exists in the spec (TO3l8). Looks like when we
added it in 228eabf we forgot to add it to the typings, which I guess is
why it then ended up being added to DeprecatedClientOptions in 4f66064.

So, add it to the public typings.

Documentation taken from sdk-api-reference at commit a8fdd42 (see [1]).

Part of #1677.

[1] https://github.com/ably/sdk-api-reference/pull/39