Add a deprecation warning for the `headers` client option
Along with an explanation of why it’s in DeprecatedClientOptions.

We’ve removed this client option in commit d7aaf34 on the integration/v2
branch. It sounds unlikely this will affect anyone, but add a
deprecation warning just in case, and to be consistent with our approach
to other deprecated client options.

Part of #1666.