Improve errors for incorrect constructor args
This fixes the regression that I introduced in e8c3012, which stopped
the user from getting a useful error message when not passing any args
to the constructor. I’ve also introduced a wider range of error messages
to give the user useful guidance depending on which variant of the
library they’re using.

Resolves #1700.