ECO-4787: Use interception proxy prototype
This adds a client for the interception proxy, and passes all test
WebSocket traffic through the proxy. It does not currently perform any
modifications to the WebSocket traffic (I’ll add a subsequent commit
where we convert a handful of tests to use the proxy instead of private
APIs for intercepting and modifying Realtime protocol messages).

This is very much a prototype and work in progress, hence all the TODOs.
The API of InterceptionProxyClient will probably evolve as we convert
more tests to use it.

I’ve also added an NPM script called test:playwright:open-browser
script, which lets you open a headed browser which is configured to use
the interception proxy. Useful for local debugging of browser tests.