Support optional multiple recovery scopes
New client option where, if specified, the SDK's recovery key will be persisted under
that identifier (in sessionstorage), so only another library instance which specifies
the same recoveryScope will attempt to recover from it. This is useful if you have
multiple ably-js instances sharing a given origin (the origin being the scope of
sessionstorage), as otherwise the multiple instances will overwrite each other's
recovery keys, and after a reload they will all try and recover the same connection,
which is not permitted and will cause broken behaviour.

renamed recoveryScope to recoveryKeyStorageName