Record usages of private APIs in the test suite
This adds annotations to each place in the test suite where a private
API is called. When the tests are run, these annotations are used to
emit a JSON file containing private API usage information. (In ECO-4834,
we’ll process this JSON file to generate some useful reports.) This JSON
file also contains a list of all of the tests which get run, which will
be used as part of the aforementioned reporting to allow us to generate
a report of the tests that do not use any private APIs.

The motivation for this gathering this data is that we intend to try and
reuse the ably-js test suite as a unified test suite for all of our
client libraries. In order to do this, we’ll need to address the usage
of ably-js private APIs in the test suite, and to do that we need to
first gather data about the current usage.

I originally intended to put the changes into the current commit into a
separate unified test suite branch, instead of merging them into `main`.
But I don’t really want to maintain a long-running feature branch, and
also I want to make sure that we add private API annotations to any new
tests that we add to ably-js, hence I decided to put this commit into
`main`. But maybe, once we’re done with the unified test suite, we’ll
decide to remove these annotations.

Given the context in which we’re gathering this data, in the interests
of saving time I skipped adding some annotations to some files that I’m
pretty sure will not form part of the unified test suite:

- test/realtime/transports.test.js
- test/browser/simple.test.js
- test/browser/http.test.js
- test/browser/connection.test.js
- test/browser/modular.test.js
- test/browser/push.test.js
- test/rest/bufferutils.test.js

If we decide at some point that we’d like to for whatever reason gather
data on these tests, we can do that later.

(Note I haven't been very consistent with how many times per test I
record the usage of a given private API; it’s not a stat I’m
particularly interested in here. But for the most part I’m doing it at
the point of each invocation of the private API.)

Resolves ECO-4821.