Record usages of private APIs in the test suite
For ECO-4821 (getting better information on private API usage, in the
context of reusing this test suite as a unified test suite).

(Note I haven't been very consistent with how many times per test I
record the usage of a given private API; it’s not a stat I’m
particularly interested in here. But for the most part I’m doing it at
the point of each invocation of the private API.)