Convert shared helper to a class
Motivation as in 9e28603. To expand a little, in order to implement
ECO-4821 (gathering detailed information about private API usage in the
test suite), we will give each test function access to a unique helper
object which is aware of the identity of the test case in which it is
executing. This helper object will offer a method for recording the fact
that a test function is calling a private API. By using the existing
shared_helper object, we can also record the usage of private APIs by
the various helper methods, access to which is currently funnelled
through the shared_helper.

The couple of methods that I’ve made static (randomString,
whenPromiseSettles) are ones that are very frequently called _and_ which
definitely don’t use any private APIs, so I’ve made them static so that,
when we make the aforementioned change giving each test case its own
helper object, we don’t need to pass a helper around just to enable
these methods to be called. The choice of making just these two methods
static was pretty arbitrary; once we’ve catalogued all of the private
API usage it might turn out there some other ones that can be made
static.