Convert shared helper to a class
TODO figure out what should be class method; at the moment everything is
instance except for the ones that have a .skip because that wasn't
implementable for an instance method, and also because these clearly
aren’t going to be called at a per-test level. But there may well be
other stuff (e.g. stuff that's used at a non-test level, or it might
just be that the instance doesn't always have a test)