Update tests so that each has its own helper instance
This removes the per-file SharedHelper instance and instead makes sure
that each context which will call SharedHelper instance methods (e.g.
test cases, test hooks) has a unique SharedHelper instance that is aware
of the context from which it is being called. (Motivation for this
change already explained in 18613c3.)