[UTS] Don’t use private APIs in closeAndFinish
This is part of removing some common private API usage for the unified
test suite.

It’s not clear to me why these private APIs were required. Even
`nextTick`, which had a comment saying that it is needed by
“channelattach_publish_invalid etc” — removing the `nextTick` does not
seem to affect those tests (although 4a90e75, which introduced the
`nextTick` usage, does describe it as a “race” so perhaps we need to see
if failures emerge over time).