Set shared helper on the Test object
TODO incorporate into 3a10b09. Context: when, in a later commit, I
attempted to modify `this.helper` inside an `it` to add to its
helperStack (for e.g. the tests generated by testOnAllTransports), it
seemed to have some weird global effect; i.e. the helper then became
used by other tests.

- in test, `this` is `Context { _runnable: Test, test: Test }`
- in root hook, `this` is `Context { currentTest: Test, _runnable: Hook }`

Setting it on the `Test` object instead of on `this` seems to fix it.