Refactor `Message` type to remove deletedAt and add updateSerial
Added an updateSerial to allow LWW semantics to be applied to a message resulting from an operation.
Removed the deletedAt timestamp, we will rely only on updateSerial to apply ordering of operations.
UpdateAt will still be used, but just so clients can render a simple datetime and not have to understand/parse the serial.