build: dont check submodules to build module
The only git submodule we need is ably-common, and this is only used as part of testing to create
test apps. We don't use it when building (despite it being a build check).

This check during build prevents other SDKs from depending on a branch of ably-js, as it doesn't have
the submodules, and can't init them.

This change therefore removes the submodules check when building the library, allowing SDKs to depend on branch
versions, if they want to.