Test that we can decrypt history messages
This was previously untested.

The new test in `crypto.test.js` is taken from a test Simon added
in #1923, but I wanted to separate this test from all the other work that’s
being done in that PR. I’ve also copied this new test to create a
similar test for the modular variant of the library.

Co-authored-by: Simon Woolf <simon.woolf@ably.com>