test/realtime/message: reduce subscribes to filtered channel flakes
The test waits to receive an end message on a channel then it runs the
assertions, this can cause flakes because we dont know if the event loop
had any other pending IO which could include delivering the 2 expected
filtered messages. Run the assertions in a setImmediate so that it executes
in the next event loop run allowing any queued messages to be delivered.

When running the test with logging the messages were being delivered but
the handler was being called after the assertions had ran.

RAR-656